fix(ia/bot): ensure to init bot, when there is no forker

+++ src/ai/src/ressources/Ressources.cpp
@@ -41,7 +41,7 @@ void Ressources::addRessource(std::string ressource)
     }
 }
 
-size_t Ressources::getRessource(const std::string &ressource) const
+int Ressources::getRessource(const std::string &ressource) const
 {
     auto it = ressourcesMap.find(ressource);
     if (it != ressourcesMap.end())
